pyproject.toml
[tool.poetry]
name = "testshop"
version = "0.1.0"
description = ""
authors = ["Aybek18 <kasymaliev09@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.4"
openpyxl = "^3.1.2"
django-redis = "^5.3.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
